A new "experience" branch has been created. Work has begun on the "experience" block.
html - The "experience" block has been created. Basic classes and tags were added to it. During the work, the structure of the block was slightly redesigned. The general class "buttons" was added for the interactive buttons on the site.

css/scss - The "experience" block and its classes were added to the "style.css/scss" file. The general classes section has been added and the "buttons" class has been added to it.